fix: run the overlay teardown tests where a desktop can actually be read
Wiring these into the hosted CI lane failed all four, and the failure was
right: a hosted runner has enough window station for UI Automation - which is
why a WPF fixture stages there quite happily - but no composited desktop to
read back, so a layered overlay reads as zero pixels. Asserting there fails
for a reason that says nothing about the renderer.

They now run from scripts/run-windows-e2e-ci.ps1, under the same exclusive
desktop lease as the rest of the live suite, and the run fails if they fail.
That is where every other live Windows check on this project already lives:
the live lane was removed from CI deliberately, and the suite is operator-run.

The skip message is the part that actually misled. It claimed the Test
(Windows) lane set the variable and owned executing these, and nothing did -
so four tests covering the whole renderer reported ok while asserting
nothing. It now names the script that runs them.

Each test also asserts `rendered` before waiting on a pixel. Without that, a
renderer that never started is indistinguishable from an overlay that never
painted, and the failure names the wrong one - which is exactly how the CI
run above read at first glance.

The harness moves to a sibling module to stay under the file size cap.
